+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'jam:soft-delete',
-  version: '0.1.1',
+  version: '0.2.0',
   summary: 'An easy way to add soft deletes to your Meteor app',
   git: 'https://github.com/jamauro/soft-delete',
   documentation: 'README.md'
@@ -11,6 +11,8 @@ Package.onUse(function(api) {
   api.use('ecmascript');
   api.use('mongo');
   api.use('check');
+  api.use('accounts-base');
+  api.use('jam:easy-schema@1.3.1', { weak: true });
   api.use('zodern:types@1.0.13');
   api.mainModule('soft-delete.js');
 });
@@ -18,6 +20,7 @@ Package.onUse(function(api) {
 Package.onTest(function(api) {
   api.use('ecmascript');
   api.use('tinytest');
+  api.use('accounts-password')
   api.use('jam:soft-delete');
   api.mainModule('tests.js');
 });