+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+* feat: automatically add the configured `deleted` flag when inserting
+* fix: add configured `deleted` flag when creating a user
+
 ## 0.1.1
 * fix: error in 3.0+ when creating or logging in a user